1.45 [Truce]
Valve updated some stuff with the Halloween bosses that don't matter to
me, but they did also add an animation for quickjoining with
Merasmissions, which do matter to me, so it's been updated.

When I say something doesn't matter, I mean that I don't need to update
anything for it. I don't mean that I don't care about the feature. Just
by the way.