1.70 [We back bois]
So this is a huge update. I wanted to tie it in with the update that we
thought was gonna hit on Thursday but obviously did not.
I also wanted to make even more changes, but I decided that enough weas
enough for this time.
There might be another big update some time in the future, but I'm happy
with how this turned out.