1.74 [Menu]
So many changes I can't list all of them.
BUT!
Got the menu back. The stock design might need some teaking to be
completely stable so I included the gamemenu.res file in that
customization for now.
Fixed issues with positioning the box around the Uber percentage.
Added black backgrounds to targetids and winpanels. These are now
default, but the coloured ones are still supported.
Started supporting different transparencies for the HP/ammo box,
including solid and completely invisible. The colours probably aren't
perfect, but it's a start.
Hundreds of other little tweaks or fixes.