1.80 [SMISSMASS + also a Steam Group]
This is a quick fix to update all the changed files. I WILL be posting
another update really soon that probably changes the background of the
"next map" panels. I just need to actually play it first and figure out
what needs changing and how.
See this as a hot fix, but also the next update is really minor, only
aesthetic stuff.

In other news, you can join my new Steam Group for this HUD to be
notified when I post updates.

http://steamcommunity.com/groups/biscottiHUD