1.83 [also effin' huge and even longer time coming]
So this was has been sitting 98% finished for about 4 months. I started
working on it right after the last update and then just lost the
motivation to finish it all the way. So, as I often do, I'll now release
the stuff that I did finish. It's some pretty cool stuff actually, but
it has been several months since I last worked on it, so forgive me if I
leave something out. These are the highlights as I remember them.

- Custom crosshairs.
Really cool actually. I now have proper good support for them and even
made a couple presets for you to try out.
Even more exciting, you can use the outer part of the crosshair as a
hitmarker, either by animating the colour or the alpha (transparency) of
it.
You can also use a different hitmarker, like konrwings or whatever you
desire, and leave the crosshair2 as is.
Both are turned off by default, but I recommend you at least try them,
if nothing else then for fun. The presets are pretty exaggerated but you
can always edit them, get rid of crosshair2, or use crosshair2's alpha
as a hitmarker.
To change the colours of the crosshairs, go to
biscottiHUD\resource\schemes\scheme_crosshaircolours.res. It will
require you to restart the game to see a change, though. You can also go
the biscottiHUD\scripts\hudlayout_hitmarker.res file but it will be
overwritten if you change to a different customization. Unless you
change it in customization\manual. That would allow you to use different
colours for different crosshairs too.
To keep it simple, try the Seeker preset. It's probably the most usable
and least intrusive. The others are mostly just something I made for fun
to try different things.

- Sticky count label
Similar to the uber charge label, there's now a label below the
crosshair that counts how many stickies you have it (it's invisible if
you have 0 out, don't worry). If you want it to still be there and say 0
when you don't have any out, go to huddemomanpipes_label3.res and change
the NoPipesPresent stuff to "visible" "1".
Again, similar to the uber label, you can choose different positions on
the HUD, or turn it off altogether. At the ammo, at the crosshair, and
at the charge bar. Ammo, and crosshair are on by default.
It's simple, but maybe one of my low-key favourite things in this
update. I think it was also the first thing I started working on
immediately after publishing the last update. I realised I missed an
opportunity.

- PDA Menu designs
This is what caused me to take a break. I wanted to do more condensed,
icon-based versions of the PDAs but I kind of got stuck. I knew that I
wanted to make the current design smaller, at the very least.
So I designed an icon-based disguise menu, which I am very happy with,
along with two vertical designs. They're practically identical, but I
couldn't decide between them so you get both for now.
I wanted to do the same for the Engineer but hit a wall. So for the past
few days, I've been optimising the normal design of the build/destroy
menu along with the Eureka teleport menu. I even did the Pip-Boy ones,
finally!
But, at least for now, the condensed, icon-based and vertical designs
will have to wait.

- Health you pick up
I made it not float around. Now it stands still and fades out. It also
moves with positioning now, so it follows where the HP box is.

- What's to come
Like I said, I want to do the other designs for the Engineer PDAs.
I also would very much like to copy those designs over to the Class menu
as well. It fits so nicely, but coding it is a little trickier and I
also need to design it first. I also really needed to push this update
because (HYPE HYPE HYPE!!) the Pyro update is coming! For those reasons,
these changes are on the back burner for a while, as they have been for
several months. I'll get to it within a couple of months, probably.

Oh, and I've been meaning to make a couple tutorial videos explaining
how to activate customizations, edit them and what not. It'll happen,
but then again, I've been saying that for about 6 months now. But it'll
happen.
But in the meantime, if you want to try some of the customizations, try
using the .bat files. They're hella convenient for switching between
options quickly.

I hope I didn't mess up somewhere, because again, this update is like
624 file changes.
And I know I should update the README more. I'll do that at some point
too. Peace!

PS: One of my many, many flaws is that I go into way to much details
with my patch notes. I'm basically writing this shit for my own
amusement. It took me, what, 40 minutes to write PATCH NOTES (!) that
will be irrelevant in 24 hours because that's when the next patch will
be. No-one's gonna read this. I hope womeone reads this. 'Sup to
whomever (yes, whom) reads this. I hope you have a splendid day!