1.84 [Jungle Inferno]
It's not completely done yet but it's pretty much there. It's a hotfix
at the very least. More will come once I get a chance to test it, mess
around, and design around the new menu changes.

- It's stable. I think.
- For the time being we're using a modified version of the stock main
menu.
- I added the Vita Saw organs in the same way I do all other counters
e.g. Revenge Crits, Heads.
- I still need to change the Pyro's jetpack. I imagine it's a charge bar
like Jarate or something but I haven't been able to test it yet as I
don't have one :(
- supposedly added in some more changes from the previous HUD update
that I for some reason hadn't copied over properly.
- added captions for new voice lines to prevent error messages in
console.

Speaking if the console. Did they add, like, some weird layering to the
main menu? It keeps clsing the console whenever I click it out of focus.
But then it takes two toggles to come back like it was there hidden
below something else but we just can't see it. That's mad annoying.

This is a lot of the changes but not all of them. I should have more
patches coming out soon. The jetpack is a priority, I just need to
acquire one first.