1.86a
I made a small mistake. Thank you for notifying me on GitHub even though I only noticed it just now, 16 days after you told me.
When Jungle Inferno shipped, the gas can and phlog used the same meter so they overlapped. I expected them to make a new meter file for the gas can and keep the phlog where it was, so I positioned the jetpack to be above the plhog's original position. They did something else; they made a new file for the phlog and I didn't even notice the mix-up. In short, the new phlog position overlaps the jetpack, so I've had to move the jetpack up a slot.

- moved the jetpack up slitghly and then wrote a way too long patch note for that one thing that I did.

But thank you for pointing it out so I could fix it :)