1.88 [Blue Moon]
So... sorry it's been so long. I haven't been playing TF2 and the update came out of freaking nowhere which spooked me. So it took me a while to get around to it, but it's fine now. It's been done. It didn't even take long. I shouldn't have freaked out about it.

- updated to be compatible with latest game update
- Valve now lets you change the colour of the damage numbers in-game now, which is good, but I cannot lie, it upset me a bit because it takes purpose away from custom HUDs. But it's a good change. I've left the colour code in just in case they change it back for whatever reason. It doesn't seem to do anything anyway.
- I haven't the foggiest what the change to hudobjectivetimepanel is supposed to accomplish but it should be fine leaving it as is. I'll update if I find any issues.